rp2040

+++ src/AudioMetaData/MetaDataICY.h
@@ -63,7 +63,7 @@ class MetaDataICY : public AbstractMetaData {
         /// Writes the data in order to retrieve the metadata and perform the corresponding callbacks 
         virtual size_t write(const uint8_t *buffer, size_t len) override {
             if (callback!=nullptr){
-                for (int j=0;j<len;j++){
+                for (size_t j=0;j<len;j++){
                     processChar((char)buffer[j]);
                 }
             }
